Crate file

The file crate is responsible for the implementation of interaction between the engine and the file system. Its main logic centers around retrieving from/writing to disk data pages.

Modules

api

Public API for the file crate

file_catalog

A file catalog mapping file IDs to their file names

in_memory_file_manager

Defines a temporary implementation for the InMemoryFileManager